feat(rips): expose embedded FLAC cover art via FlacReader
